requirements.txt
asgiref==3.8.1
Django==5.0.6
django-cors-headers==4.4.0
django-filter==24.2
djangorestframework==3.15.2
djangorestframework-camel-case==1.4.2
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
psycopg2==2.9.9
sqlparse==0.5.0
whitenoise==6.7.0
